requirements.txt
streamlit
opencv-python
matplotlib
Pillow
streamlit-image-coordinates
numpy
torch
torchvision
torchaudio
segment_anything